StageObject zSoringOffset added for objects with Z depth, All objects face right by default
The StageObject CompareTo function still needs tweaking. It works
perfect for right facing objects when another object is standing on top
or is on the left side. But when on the right side (when not not
standing on it), the sorting should be reversed. This entire process is
the opposite for left facing objects.